Voronoi displaying on sloped plane should have been done
I do not know, how the result should look like, but imo the voronoi diagram on slope should work. The voronoi diagram is displayed on plane defined by the basePoint and the normal vector of plane. Currently, I am not sampling some function - point are random generated. Points are projected on the plane afterwards. Projected points are displayed on the ground (z = 0). This process is made with rotation around arbitrary axis. The axis is defined as a intersection of the point plane with ground plane. Points are rotated on the ground and voronoi diagram is calculated then. Lastly, voronoi cells of diagram are projected reversively to the plane and drawn on the canvas.